moved handler functions to own files in like directories
